refactor: 🚚 Register ppc-update-shipping unconditionally in ppcp-order-endpoints
