+++ modules/ppcp-blocks/src/BlocksModule.php
@@ -11,7 +11,6 @@ namespace WooCommerce\PayPalCommerce\Blocks;
 
 use Automattic\WooCommerce\Blocks\Payments\PaymentMethodRegistry;
 use WooCommerce\PayPalCommerce\Assets\AssetGetter;
-use WooCommerce\PayPalCommerce\Blocks\Endpoint\UpdateShippingEndpoint;
 use WooCommerce\PayPalCommerce\Button\Assets\SmartButtonInterface;
 use WooCommerce\PayPalCommerce\Vendor\Inpsyde\Modularity\Module\ExecutableModule;
 use WooCommerce\PayPalCommerce\Vendor\Inpsyde\Modularity\Module\ExtendingModule;
@@ -86,16 +85,6 @@ class BlocksModule implements ServiceModule, ExtendingModule, ExecutableModule {
 					return array();
 				},
 			)
-		);
-
-		add_action(
-			'wc_ajax_' . UpdateShippingEndpoint::ENDPOINT,
-			static function () use ( $c ) {
-				$endpoint = $c->get( 'blocks.endpoint.update-shipping' );
-				assert( $endpoint instanceof UpdateShippingEndpoint );
-
-				$endpoint->handle_request();
-			}
 		);
 
 		// Enqueue frontend scripts.
