+++ modules/ppcp-order-endpoints/src/OrderEndpointsModule.php
@@ -15,6 +15,7 @@ namespace WooCommerce\PayPalCommerce\OrderEndpoints;
 use WooCommerce\PayPalCommerce\OrderEndpoints\Endpoint\ApproveOrderEndpoint;
 use WooCommerce\PayPalCommerce\OrderEndpoints\Endpoint\ChangeCartEndpoint;
 use WooCommerce\PayPalCommerce\OrderEndpoints\Endpoint\CreateOrderEndpoint;
+use WooCommerce\PayPalCommerce\OrderEndpoints\Endpoint\UpdateShippingEndpoint;
 use WooCommerce\PayPalCommerce\Vendor\Inpsyde\Modularity\Module\ExecutableModule;
 use WooCommerce\PayPalCommerce\Vendor\Inpsyde\Modularity\Module\ModuleClassNameIdTrait;
 use WooCommerce\PayPalCommerce\Vendor\Inpsyde\Modularity\Module\ServiceModule;
@@ -72,6 +73,16 @@ class OrderEndpointsModule implements ServiceModule, ExecutableModule {
 				 *
 				 * @var CreateOrderEndpoint $endpoint
 				 */
+				$endpoint->handle_request();
+			}
+		);
+
+		add_action(
+			'wc_ajax_' . UpdateShippingEndpoint::ENDPOINT,
+			static function () use ( $c ) {
+				$endpoint = $c->get( 'order-endpoints.endpoint.update-shipping' );
+				assert( $endpoint instanceof UpdateShippingEndpoint );
+
 				$endpoint->handle_request();
 			}
 		);
